update to 1.2.6

+++ src/linux.rs
@@ -16,12 +16,17 @@ pub fn update(base_url: &str, version: &str, temp_dir: PathBuf) -> () {
 
 fn linux_arch_download(url: String, temp_path: &str) {
     download_from_url(url, temp_path);
-    Command::new("sudo")
+    let output = Command::new("sudo")
         .arg("pacman")
         .arg("-U")
-        .arg(temp_path)
         .arg("--noconfirm")
+        .arg(temp_path)
         .output()
         .expect("Failed to install");
+
+    if !output.status.success() {
+        panic!("{}", String::from_utf8_lossy(&output.stderr));
+    }
+
     remove_file(temp_path).expect("Failed to remove temp file");
 }
